requirements.txt
flask
flask-pymongo
python-slugify
python-dotenv
